Se agrega funcion que valida si la extension del archivo es correcta
